Included the vector.cpp file
